+++ test/test-80-compression-node-opcua/main.js
@@ -12,6 +12,10 @@ const fs = require('fs');
 const path = require('path');
 const assert = require('assert');
 const utils = require('../utils.js');
+const pkgJson = require('./package.json');
+
+const isWindows = process.platform === 'win32';
+const buildDir = 'build';
 
 assert(!module.parent);
 assert(__dirname === process.cwd());
@@ -20,18 +24,23 @@ if (utils.shouldSkipPnpm()) {
   return;
 }
 
+function clean() {
+  utils.vacuum.sync(buildDir);
+  utils.vacuum.sync('node_modules');
+  utils.vacuum.sync('./pnpm-lock.yaml');
+}
+
 // remove any possible left-over
-utils.vacuum.sync('./node_modules');
-utils.vacuum.sync('./pnpm-lock.yaml');
+clean();
+
+const npmlog = utils.exec.sync('npm install -g pnpm@8');
+console.log('npm log :', npmlog);
 
 // launch `pnpm install`
 const pnpmlog = utils.spawn.sync(
-  path.join(
-    path.dirname(process.argv[0]),
-    'npx' + (process.platform === 'win32' ? '.cmd' : ''),
-  ),
+  path.join(path.dirname(process.argv[0]), 'npx' + (isWindows ? '.cmd' : '')),
   ['pnpm', 'install'],
-  { cwd: path.dirname(__filename), expect: 0 },
+  { cwd: path.dirname(__filename), expect: 0, shell: isWindows },
 );
 console.log('pnpm log :', pnpmlog);
 
@@ -46,12 +55,12 @@ assert(
 /* eslint-disable no-unused-vars */
 const input = 'package.json';
 const target = process.argv[2] || 'host';
-const ext = process.platform === 'win32' ? '.exe' : '';
-const outputRef = 'test-output-empty' + ext;
-const outputNone = 'test-output-None' + ext;
-const outputGZip = 'test-output-GZip' + ext;
-const outputBrotli = 'test-output-Brotli' + ext;
-const outputBrotliDebug = 'test-output-Brotli-debug' + ext;
+const ext = isWindows ? '.exe' : '';
+const outputRef = path.join(buildDir, 'test-output-empty' + ext);
+const outputNone = path.join(buildDir, 'test-output-None' + ext);
+const outputGZip = path.join(buildDir, 'test-output-GZip' + ext);
+const outputBrotli = path.join(buildDir, 'test-output-Brotli' + ext);
+const outputBrotliDebug = path.join(buildDir, 'test-output-Brotli-debug' + ext);
 
 const inspect = ['ignore', 'ignore', 'pipe'];
 
@@ -78,12 +87,36 @@ function pkgCompress(compressMode, output) {
   );
   // check that produced executable is running and produce the expected output.
   const log = utils.spawn.sync(path.join(__dirname, output), [], {
-    cwd: __dirname,
+    cwd: path.join(__dirname, buildDir),
     expect: 0,
   });
   assert(log === '42\n');
   return fs.statSync(output).size;
 }
+
+function esbuildBuild(entryPoint) {
+  const log = utils.spawn.sync(
+    path.join(path.dirname(process.argv[0]), 'npx' + (isWindows ? '.cmd' : '')),
+    [
+      'esbuild',
+      entryPoint,
+      '--bundle',
+      '--outfile=' + path.join(buildDir, pkgJson.main),
+      '--platform=node',
+    ],
+    { cwd: __dirname, expect: 0, shell: isWindows },
+  );
+
+  console.log(log);
+
+  // copy folder 'node_modules/node-opcua-nodesets' to build folder
+  utils.copyRecursiveSync(
+    'node_modules/node-opcua-nodesets/nodesets',
+    path.join(buildDir, 'nodesets'),
+  );
+}
+
+esbuildBuild(pkgJson.main);
 
 const sizeNoneFull = pkgCompress('None', outputNone);
 const sizeGZipFull = pkgCompress('GZip', outputGZip);
@@ -121,15 +154,8 @@ const logPkg5 = utils.pkg.sync(
   { expect: 2 },
 );
 
-// xx console.log(logPkg4);
 assert(logPkg5.match(/Invalid compression algorithm/g));
 
-utils.vacuum.sync(outputRef);
-utils.vacuum.sync(outputNone);
-utils.vacuum.sync(outputBrotli);
-utils.vacuum.sync(outputGZip);
-utils.vacuum.sync(outputBrotliDebug);
-utils.vacuum.sync('node_modules');
-utils.vacuum.sync('./pnpm-lock.yaml');
+clean();
 
 console.log('OK');